move litentry/litmus/moonbase/rococo transaction_payment.rs to runtime_common/lib.rs:
